Extract models to DevnetClientModels
